added persistence option to file download servlet, added simple epc info to prov wizard

+++ TigerSpice_dev/src/com/velisphere/tigerspice/server/DownloadServlet.java
@@ -45,7 +45,12 @@ public class DownloadServlet extends HttpServlet{
 	        } finally {
 	        	os.close();
 	        }
-	        downloadableFile.delete();
+	        
+	        String persist = req.getParameter("persist");
+	        if (persist.equals("1") == false){
+	        	downloadableFile.delete();	
+	        }
+	        
 
     }
 }
